+++ src/types/FabloConfigJson.ts
@@ -13,6 +13,7 @@ export interface OrganizationDetailsJson {
 
 export interface CAJson {
   prefix: string;
+  db: "sqlite" | "postgres";
 }
 
 export interface OrdererJson {