+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Generate connection profiles for Node and Java gateways
 * Create a full network state snapshot in tag.gz file and restore it
 * Add [Hyperledger Explorer](https://github.com/hyperledger/blockchain-explorer) support
+* Support postgres database for CA
 
 ### Fixes